Lab 3 fixed cutting

+++ OOP_lab3_WAV/WavModifyer.hpp
@@ -26,7 +26,7 @@ private:
     void attachReverb(size_t samplesCountPerChan, int delaySamples, float decay, std::vector<float> signal);
     void scaleToShort(size_t i, size_t samplesCountPerChan, float maxMagnitude, std::vector<float> signal, DataArray &data);
     float findMaxMagnitude(size_t samplesCountPerChan, int delaySamples, std::vector<float> signal);
-    unsigned short convertSecondsToBytes(float secs, int bitsPerSample, int sampleRate) const;
+    int convertSecondsToBytes(float secs, int bitsPerSample, int sampleRate) const;
     WavWriter writer;
     ModifyState currState;
 };
